Crate rusoto_apigatewaymanagementapi

The Amazon API Gateway Management API allows you to directly manage runtime aspects of your deployed APIs. To use it, you must explicitly set the SDK's endpoint to point to the endpoint of your deployed API. The endpoint will be of the form https://{api-id}.execute-api.{region}.amazonaws.com/{stage}, or will be the endpoint corresponding to your API's custom domain and base path, if applicable.

If you're using the service, you're probably looking for ApiGatewayManagementApiClient and ApiGatewayManagementApi.

Structs

ApiGatewayManagementApiClient
A client for the AmazonApiGatewayManagementApi API.

PostToConnectionRequest

Enums

PostToConnectionError
Errors returned by PostToConnection

Traits

ApiGatewayManagementApi
Trait representing the capabilities of the AmazonApiGatewayManagementApi API. AmazonApiGatewayManagementApi clients implement this trait.